requirements.txt
# torch==2.3.1
fastapi==0.109.0
uvicorn==0.27.0
langchain==0.1.0
langchain-community==0.0.20
langchain-google-genai==0.0.11
pymongo==4.6.0
transformers==4.36.0
rank-bm25==0.2.2
scikit-learn==1.3.2
streamlit==1.31.0
python-dotenv==1.0.0
nest-asyncio==1.5.8
python-multipart==0.0.20
pypdf==5.3.1
chromadb==0.6.3
